Rename some files to fit the rest of the project and refactor some code

+++ src/main/java/com/example/demo/service/RecipeService.java
@@ -52,6 +52,7 @@ public class RecipeService {
     }
 
     public void deleteRecipe(Long id) {
+        if(getRecipe(id) == null) throw new RecipeNotFoundException(id);
         repository.deleteById(id);
     }
 }
